src/osprey/templates/services/ariel-web/requirements.txt
# ARIEL Web Interface dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0

# Database (via osprey)
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# Note: osprey is mounted from host, not installed via pip
